@@ -1,39 +1,40 @@
 module github.com/speedata/xts
 
-go 1.21
+go 1.24.0
 
 require (
 	github.com/alecthomas/chroma v0.10.0
-	github.com/boxesandglue/baseline-pdf v0.0.0-20240805104743-f940543238d5
-	github.com/boxesandglue/boxesandglue v0.0.0-20240805104743-8b6877b08561
-	github.com/boxesandglue/csshtml v0.0.0-20240813134317-f214983f48f8
-	github.com/boxesandglue/htmlbag v0.0.0-20240813191957-a1a46993bb22
-	github.com/boxesandglue/textlayout v0.0.0-20240805104743-4ef76f4729dd
+	github.com/boxesandglue/baseline-pdf v1.0.11
+	github.com/boxesandglue/boxesandglue v0.1.13
+	github.com/boxesandglue/csshtml v0.0.1
+	github.com/boxesandglue/htmlbag v0.0.2
+	github.com/boxesandglue/textlayout v1.0.5
 	github.com/cjoudrey/gluahttp v0.0.0-20201111170219-25003d9adfa9
 	github.com/gammazero/workerpool v1.1.3
 	github.com/mitchellh/mapstructure v1.5.0
 	github.com/pelletier/go-toml/v2 v2.0.8
 	github.com/speedata/barcode v0.0.0-20230307084753-486c4d51cb34
 	github.com/speedata/goxlsx v1.0.2
-	github.com/speedata/goxml v0.0.0-20230322084327-216af189a791
+	github.com/speedata/goxml v1.0.5
 	github.com/speedata/goxpath v0.0.0-20230709190058-ca75de231e9c
-	github.com/speedata/optionparser v1.0.2
+	github.com/speedata/optionparser v1.1.1
 	github.com/yuin/goldmark v1.4.13
 	github.com/yuin/goldmark-highlighting v0.0.0-20220208100518-594be1970594
 	github.com/yuin/gopher-lua v1.1.0
-	golang.org/x/net v0.28.0
-	golang.org/x/sys v0.23.0
-	golang.org/x/text v0.17.0
+	golang.org/x/net v0.48.0
+	golang.org/x/sys v0.39.0
+	golang.org/x/text v0.32.0
 )
 
 require (
-	github.com/PuerkitoBio/goquery v1.9.2 // indirect
-	github.com/andybalholm/cascadia v1.3.2 // indirect
-	github.com/boxesandglue/gofpdi v1.0.19 // indirect
+	github.com/PuerkitoBio/goquery v1.11.0 // indirect
+	github.com/andybalholm/cascadia v1.3.3 // indirect
+	github.com/beevik/etree v1.5.0 // indirect
+	github.com/boxesandglue/gofpdi v1.0.21 // indirect
 	github.com/dlclark/regexp2 v1.4.0 // indirect
 	github.com/gammazero/deque v0.2.0 // indirect
 	github.com/google/uuid v1.6.0 // indirect
-	github.com/speedata/css/scanner v0.0.0-20240704081300-00e8c453c7d5 // indirect
+	github.com/speedata/css/scanner v0.0.0-20250825095519-68d469dc62c4 // indirect
 	github.com/speedata/hyphenation v1.0.2 // indirect
-	golang.org/x/image v0.19.0 // indirect
+	golang.org/x/image v0.27.0 // indirect
 )
